Move tests to a separate folder
