+++ test/fixtures/multiArgWorker.js
@@ -1,4 +1,4 @@
-const { runAsWorker } = require('../index.js')
+const { runAsWorker } = require('../../src/index.js')
 
 runAsWorker(async (...args) => {
   return args.length
