+++ test/index.test.js
@@ -1,7 +1,7 @@
 const { it } = require('node:test')
 const assert = require('node:assert')
 
-const { createSyncFn } = require('./index.js')
+const { createSyncFn } = require('../src/index.js')
 
 it('should receive a result synchronously', () => {
   const mySyncFn = createSyncFn(require.resolve('./fixtures/worker.js'))
